docs: Update Blueprints module README and enhance mapping logic
- Added prerequisites section to the README for the Blueprints module, detailing required services for configuration.
- Updated mapping profiles in EntityDefinitionStateMappingProfile and EntityDefinitionMappingProfile to utilize FieldDefinitionMapper for improved clarity and maintainability.
- Enhanced delete operation in EntityDefinitionRepository to ensure proper handling of non-deleted entities.
- Improved integration tests for entity definition endpoints to validate success responses on creation and deletion operations.
